Initial version of a steganography library.
This provides a base class and a couple of subclasses showing simple
steganographic encoding of text into images.  There are some functional
tests that encode the play Macbeth into an image and verify that it is
recoverable.  Some refactoring and changes should be made to simplify
creation of unit tests.  Those unit tests should also be written.